Update issue 15 Thanks for your report!
It seems to function correctly for me. Does it happen on all pages or just a specific one? If it happens on all pages, I probably have to file a support ticket as that would probably mean something is broken on the website. 

Also, could you please test if the following link still functions correctly (it should send an email to 'test@test.test' with subject 'test' and body 'test')?
http://ssl.fastmail.fm/action/compose/?to=test%40test.test&cc=&bcc=&subject=test&body=test


I've updated the text to "Mail this page's URL". I didn't use 'Send/mail link' so it can't be confused with hyperlinks. The next version will contain the new text. (I'll probably release the next version when Chrome 20 is stable, or when I make code changes). Thanks for the suggestion! As a non-native English speaker (I'm Dutch), I sometimes make mistakes :)